Allow filtering messages by moderation status
I had to add some more code to handle the case where `moderation_status` is
None (i.e. message is pending moderation). As the filters use `None` to also
encode "no filter, show all", I had to change the code to use the string
"pending" to describe `moderation_status IS NULL`. It was a bit tricky to be
able to do so, without repeating the moderation status label names, but I found
an OK-ish solution.