Extract Esic model from PublicBody and add municipality and UF
A single eSIC (e.g. the CGU's one) can contain hundreds of public bodies.
Instead of repeating it, we use a separate table. This allow us to add the
"username" to the eSIC, which is the account used to create FOIA requests in
this eSIC. Note that there's no password. This is intentional, as we can use
softwares like LastPass to share the passwords securely inside a group.