Allow setting the Message receiver in admin
This is required to create a FOIRequest from the FOIRequest admin page.
However, it does makes the UI more complicated, because you can't set both the
`sender` and `receiver`. We should probably remove them and leave just
`receiver` (with a different name), that's when None means that it's the user.

I added a validation to ensure a Message can't have both sender and receiver,
while we don't change this.