Add FOIRequest.can_publish and refactor new FOIRequest form
The old "message_new" URL now is in "foi_request_new" and has two forms, one
for the FOIRequest and one for the Message. In the future, this change will
allow us to remove the code that automatically creates a FOIRequest for a
Message if it doesn't have one, which will clean the code a bit.