Add FOIRequest.previous_protocol to point to a FOIA's previous request
Sometimes a FOIA request is a continuation of a previous one. This new field
allows the user to point to the previous FOIA request's protocol.

This is a simple CharField instead of an actual Foreing Key because we don't
want to raise an error if the user adds an inexistent protocol to avoid leaking
information on what protocols exist.